requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
cffi==1.17.1
cryptography==44.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
dgl==2.4.0+cu121
docopt==0.6.2
ecdsa==0.19.0
et-xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.2.0
ftpretty==0.4.0
fuzzywuzzy==0.18.0
goatools==1.4.12
h11==0.14.0
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
Levenshtein==0.25.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
packaging==24.2
pandas==1.3.5
passlib==1.7.4
patsy==1.0.1
pillow==10.4.0
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydot==3.0.4
pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-Levenshtein==0.25.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rapidfuzz==3.9.7
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
statsmodels==0.14.1
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.0
tqdm==4.67.1
triton==3.0.0
TxGNN==0.0.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.33.0
XlsxWriter==3.2.2
zipp==3.20.2